refactored forms files
