+++ .env
@@ -6,3 +6,4 @@ VITE_FIREBASE_MESSAGING_SENDER_ID=521972181449
 VITE_FIREBASE_APP_ID=1:521972181449:web:c2abc65bc48d428e79a018
 VITE_FIREBASE_MEASUREMENT_ID=G-19WM8LF324
 REACT_APP_FIREBASE_STORAGE_BUCKET=sushi-admin.firebasestorage.app
+VITE_STRIPE_PUBLIC_KEY=pk_test_51SKLy5JCjhZTXXNg1btNxc7ydabri7jgxeUzWWVxk6GznP7EqjyHnnshmWths9dI0N0lacZDIl6HkAWHuKGzupk900Fl0JU2S2
