+++ index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link href="./src/index.css" rel="stylesheet" />
-    <title>shushi-admin</title>
+    <title>Pacifique - CEVICHE | SUSHI BAR</title>
   </head>
   <body>
     <div id="root"></div>
